# app/routers/agent.py
"""Agent control-plane routes: overview (config + queue + cost + history) and
on-demand per-ticker analysis. The Agent page is the single control surface; edits
to the agent config go through /api/settings/agents (see system.py)."""
from __future__ import annotations

from typing import Any

from fastapi import APIRouter, BackgroundTasks, HTTPException, Request

from app.actions.agents import AgentActions
from app import job_control, panel_snapshot
from app.contracts import AgentAnalyzeInput
from app.data_access import config as config_owner
from app.data_access import loaders
from app.request_security import require_local_request
from app.response_contracts import (
    AgentAnalyzeResponse,
    AgentExperimentResponse,
    AgentOverviewResponse,
    AgentResearchPromptResponse,
)
from investment_panel.core.daily_research_prompt import build_daily_research_prompt

router = APIRouter()


def _actions() -> AgentActions:
    # Route actions must share the app's request-config seam so test and
    # alternate-runtime callers never fall through to the host's live database.
    return AgentActions(config_owner.load_config(), job_control.start_refresh_job)


@router.get("/api/agent", response_model=AgentOverviewResponse, response_model_exclude_unset=True)
def agent_overview() -> dict[str, Any]:
    return _actions().overview()


@router.get("/api/agent/experiments/current", response_model=AgentExperimentResponse, response_model_exclude_unset=True)
def current_agent_experiment() -> dict[str, Any]:
    """Expose only the experiment conclusion and gates, never batch rows."""

    return _actions().current_experiment()


@router.get("/api/agent/research-prompt", response_model=AgentResearchPromptResponse, response_model_exclude_unset=True)
def agent_research_prompt() -> dict[str, Any]:
    _, research_data = panel_snapshot.context(
        cache_key="agent:daily-research",
        loader=loaders.load_daily_research_panel_data,
    )
    research_prompt = build_daily_research_prompt(
        research_data.tables,
        status={
            "ready": research_data.status.ready,
            "message": research_data.status.message,
            "source": research_data.status.source,
            "metadata": research_data.metadata,
        },
    )
    return research_prompt


@router.post("/api/agent/analyze", response_model=AgentAnalyzeResponse, response_model_exclude_unset=True)
def analyze_ticker(payload: AgentAnalyzeInput, request: Request, background_tasks: BackgroundTasks) -> dict[str, Any]:
    require_local_request(request)
    actions = _actions()
    try:
        result = actions.queue_analysis(payload.ticker, prompt=payload.prompt or "")
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc
    job = result["job"]
    if job.get("created"):
        background_tasks.add_task(
            job_control.execute_background_refresh_job,
            job["id"],
            "run_option_agents_ondemand",
            actions.config.database.url,
        )
    panel_snapshot.invalidate_context_cache()
    return result