# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
      "AgentAnalyzeInput": {
        "properties": {
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker"
        ],
        "title": "AgentAnalyzeInput",
        "type": "object"
      },
      "AgentAnalyzeResponse": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/RefreshJobResponse"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "request_id",
          "job"
        ],
        "title": "AgentAnalyzeResponse",
        "type": "object"
      },
...
      "ApiStatusResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "ready",
          "message",
          "source"
        ],
        "title": "ApiStatusResponse",
        "type": "object"
      },
...
      "EventScoutEventsResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApiStatusResponse"
          },
          "tables": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TablePayloadResponse"
            },
            "title": "Tables",
            "type": "object"
          }
        },
        "required": [
          "status",
          "tables"
        ],
        "title": "EventScoutEventsResponse",
        "type": "object"
      },
      "EventScoutPacketsResponse": {
        "properties": {
          "count": {
            "default": 0,
            "title": "Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatusResponse"
          }
        },
        "required": [
          "status"
        ],
        "title": "EventScoutPacketsResponse",
        "type": "object"
      },
      "EventScoutReplayResponse": {
        "additionalProperties": true,
        "properties": {
          "packet": {
            "additionalProperties": true,
            "title": "Packet",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "packet"
        ],
        "title": "EventScoutReplayResponse",
        "type": "object"
      },
      "EventScoutSignalInput": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "event_kind": {
            "default": "material_event",
            "maxLength": 96,
            "minLength": 1,
            "title": "Event Kind",
            "type": "string"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "observed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "source_kind": {
            "default": "unknown",
            "title": "Source Kind",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "symbol": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "trigger_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Trigger Type",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "trigger_type"
        ],
        "title": "EventScoutSignalInput",
        "type": "object"
      },
      "EventScoutSignalResponse": {
        "additionalProperties": true,
        "properties": {
          "signal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "EventScoutSignalResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefreshJobResponse": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "job_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Name"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "summary": {
            "title": "Summary"
          }
        },
        "title": "RefreshJobResponse",
        "type": "object"
      },
...
      "TablePayloadResponse": {
        "properties": {
          "count": {
            "default": 0,
            "title": "Count",
            "type": "integer"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiStatusResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TablePayloadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Personal Investment Panel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/agent/analyze": {
      "post": {
        "operationId": "analyze_ticker_api_agent_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentAnalyzeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Ticker"
      }
    },
...
    "/api/event-scout": {
      "get": {
        "operationId": "event_scout_events_api_event_scout_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScoutEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Scout Events"
      }
    },
    "/api/event-scout/packets": {
      "get": {
        "operationId": "event_scout_packets_api_event_scout_packets_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScoutPacketsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Scout Packets"
      }
    },
    "/api/event-scout/replay": {
      "get": {
        "operationId": "event_scout_replay_api_event_scout_replay_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "default": "MRNA",
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScoutReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Scout Replay"
      },
      "post": {
        "operationId": "persist_event_scout_replay_api_event_scout_replay_post",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "default": "MRNA",
              "maxLength": 16,
              "minLength": 1,
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScoutReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Persist Event Scout Replay"
      }
    },
    "/api/event-scout/signals": {
      "post": {
        "operationId": "event_scout_signal_api_event_scout_signals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventScoutSignalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventScoutSignalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Scout Signal"
      }
    },
...
  }
}